requirements.txt
GPUtil==1.4.0
GPUtil==1.4.0
numpy==1.24.2
opencv_python==4.9.0.80
packaging==24.2
Pillow==11.1.0
setuptools==75.8.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
xarray==2024.7.0
ConfigArgParse==1.5.3